requirements.txt
python-dotenv==1.0.0
praw==7.7.1
Faker==22.6.0